wip small refactor and documentation
