+++ packages/blueprint/src/codecs.ts
@@ -30,4 +30,6 @@ const StringSplitCodec = new t.Type<string, StringUnder64[], StringUnder64[]>(
   }
 );
 
-export const StringCodec = t.array(StringUnder64Guard).pipe(StringSplitCodec);
+export const StringCodec: t.Type<string, string[], unknown> = t
+  .array(StringUnder64Guard)
+  .pipe(StringSplitCodec);
