+++ packages/blueprint/src/index.ts
@@ -1,1 +1,21 @@
-export { Blueprint, mkBlueprint, MkBlueprint, BlueprintOf } from "./typed.js";
+/**
+ * Something something
+ * @packageDocumentation
+ */
+export {
+  Blueprint,
+  mkBlueprint,
+  MkBlueprintOptions,
+  BlueprintOf,
+  Expand,
+} from "./blueprint.js";
+export {
+  BlueprintParam,
+  BlueprintType,
+  StringParam,
+  ValueParam,
+  AddressParam,
+  DateParam,
+  BlueprintKeys,
+  TypeOfParam,
+} from "./blueprint-param.js";
